Make RouteObserver generic
