+++ packages/flutter/lib/src/widgets/pages.dart
@@ -128,85 +128,3 @@ class PageRouteBuilder<T> extends PageRoute<T> {
   }
 
 }
-
-/// A [Navigator] observer that notifies [PageRouteAware]s of changes to the
-/// state of their [PageRoute].
-///
-/// To make a [StatefulWidget] aware of its current [PageRoute] state, implement
-/// [PageRouteAware] in its [State] and subscribe it to the [PageRouteObserver]:
-///
-/// ```dart
-/// class PageRouteAwareWidgetState extends State<PageRouteAwareWidget> with PageRouteAware {
-///   @override
-///   void didChangeDependencies() {
-///     super.didChangeDependencies();
-///     pageRouteObserver.subscribe(this, ModalRoute.of(context));
-///   }
-///
-///   @override
-///   void dispose() {
-///     pageRouteObserver.unsubscribe(this);
-///     super.dispose();
-///   }
-///
-///   @override
-///   void didPush() {
-///     // Do something
-///   }
-///
-///   @override
-///   void didPopNext() {
-///     // Do something
-///   }
-///
-/// }
-/// ```
-class PageRouteObserver extends NavigatorObserver {
-  final Map<PageRoute<dynamic>, PageRouteAware> _listeners = <PageRoute<dynamic>, PageRouteAware>{};
-
-  void subscribe(PageRouteAware routeAware, PageRoute<dynamic> route) {
-    assert(routeAware != null);
-    assert(route != null);
-    if (!_listeners.containsKey(route)) {
-      routeAware.didPush();
-      _listeners[route] = routeAware;
-    }
-  }
-
-  void unsubscribe(PageRouteAware routeAware) {
-    assert(routeAware != null);
-    _listeners.remove(routeAware);
-  }
-
-  @override
-  void didPop(Route<dynamic> route, Route<dynamic> previousRoute) {
-    if (route is PageRoute && previousRoute is PageRoute) {
-      _listeners[previousRoute]?.didPopNext();
-      _listeners[route]?.didPop();
-    }
-  }
-
-  @override
-  void didPush(Route<dynamic> route, Route<dynamic> previousRoute) {
-    if (route is PageRoute && previousRoute is PageRoute) {
-      _listeners[previousRoute]?.didPushNext();
-    }
-  }
-}
-
-/// A interface that is aware of its current PageRoute.
-abstract class PageRouteAware {
-  /// Called when the top route has been popped off, and the current route
-  /// shows up.
-  void didPopNext() { }
-
-  /// Called when the current route has been pushed.
-  void didPush() { }
-
-  /// Called when the current route has been popped off.
-  void didPop() { }
-
-  /// Called when a new route has been pushed, and the current route is no
-  /// longer visible.
-  void didPushNext() { }
-}
