+++ packages/flutter/test/widgets/routes_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:collection';
 
+import 'package:mockito/mockito.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/widgets.dart';
 
@@ -397,4 +398,45 @@ void main() {
       ]
     );
   });
+
+  group('PageRouteObserver', () {
+    test('calls correct listeners', () {
+      final RouteObserver<PageRoute<dynamic>> observer = new RouteObserver<PageRoute<dynamic>>();
+      final RouteAware pageRouteAware1 = new MockRouteAware();
+      final MockPageRoute route1 = new MockPageRoute();
+      observer.subscribe(pageRouteAware1, route1);
+      verify(pageRouteAware1.didPush()).called(1);
+
+      final RouteAware pageRouteAware2 = new MockRouteAware();
+      final MockPageRoute route2 = new MockPageRoute();
+      observer.didPush(route2, route1);
+      verify(pageRouteAware1.didPushNext()).called(1);
+
+      observer.subscribe(pageRouteAware2, route2);
+      verify(pageRouteAware2.didPush()).called(1);
+
+      observer.didPop(route2, route1);
+      verify(pageRouteAware2.didPop()).called(1);
+      verify(pageRouteAware1.didPopNext()).called(1);
+    });
+
+    test('does not call listeners for non-PageRoute', () {
+      final RouteObserver<PageRoute<dynamic>> observer = new RouteObserver<PageRoute<dynamic>>();
+      final RouteAware pageRouteAware = new MockRouteAware();
+      final MockPageRoute pageRoute = new MockPageRoute();
+      final MockRoute route = new MockRoute();
+      observer.subscribe(pageRouteAware, pageRoute);
+      verify(pageRouteAware.didPush());
+
+      observer.didPush(route, pageRoute);
+      observer.didPop(route, pageRoute);
+      verifyNoMoreInteractions(pageRouteAware);
+    });
+  });
 }
+
+class MockPageRoute extends Mock implements PageRoute<dynamic> { }
+
+class MockRoute extends Mock implements Route<dynamic> { }
+
+class MockRouteAware extends Mock implements RouteAware { }
